--- frontend/react-practice-ts/src/pages/approvalTempPage.tsx
@@ -1,26 +1,0 @@
-
-import { ApprovalHeader } from '../components/approval/approvalHeader';
-import { ApprovalSearchBar } from '../components/approval/approvalSearchBar';
-import { ApprovalTempBody } from '../components/approval/approvalTempBody';
-import { ApprovalTempFooter } from '../components/approval/approvalTempFooter';
-import Header from '../components/common/Header';
-import Sidebar from '../components/common/Sidebar';
-
-
-export const ApprovalTempPage = () => {
-
-  return (
-    <div className="mainpageContainer">
-    <Sidebar />
-    <div className="componentContainer">
-        <Header/>
-        <div className="componentContainer1">
-          <ApprovalHeader />
-          <ApprovalSearchBar />
-          <ApprovalTempBody />
-          <ApprovalTempFooter />
-        </div>
-      </div>
-    </div>
-  );
-}
